Fix author node handling
The occurrences of author elements is 1 or 0.
However, according to configuration algorithm,
the second occurrences of author element should be ignored.
Even if the author element appears twice, installation should not be failed.

Please reference below.
http://www.w3.org/TR/widgets/#algorithm-to-process-a-configuration-document
9.1.2 Algorithm to Process a Configuration Document
<author> element
If this is not the first content element encountered by the user agent,
then the user agent MUST ignore this element.

This issue detected from tct-widget01-w3c-tests, b7.wgt.

Change-Id: Ic14ea7683dda3b3016aefd18820245cf4dbf7b66