Provide abstraction of W3C config.xml file parser
This class will allow to add validation that requires more than
just simple checks to configuration parsing.

Source code of this class should follow:
 - http://www.w3.org/TR/widgets/#algorithm-to-process-a-configuration-document

and fix missing checks (the one that are not present in handlers at all).

This class will be aware of underlaying filesystem in the opposite to
ManifestParser class.

Change-Id: I521456b53ce1a29d2688e820e8e943bcf93b3dff