Rework ContentHandler
According to w3c spec:
 - http://www.w3.org/TR/widgets/, 9.1.2, 9, content element section.

"If this is not the first content element encountered by the user agent,
then the user agent must ignore this element."

This implies that any w3c content element that is not first must be ignored.
Therefore fixing ContentHandler.

What is more, related tests are showing that the second element will not
become the first in case of ignoring the first:
 - bq,
 - br (invalid),
 - bs,
 - d7,
 - d8,
 - d9 (invalid).

To verify:
 - run to given w3c testcases,
 - run unittests: /usr/bin/manifest-parser-ut/content_handler_unittest,
 - run widget02 automatic tests (passrate 66/68, 2 broken tests which manually works).

Change-Id: I2612cd5e5211169cc1782f97d7f70639caeee9c1