Fix LanguageTagValidator
http://www.w3.org/International/articles/language-tags/

The entries in the registry follow certain conventions with regard
to upper and lower letter-casing. For example, language tags are
lower case, alphabetic region subtags are upper case, and script
tags begin with an initial capital. This is only a convention!
When you use these subtags you are free to do as you like, unless
you are constrained by the rules of the system you are working with.
For HTML and XML language markup, the case should not matter.

Change-Id: I0e5c146c6c86b64f7d1a36bd76cae9b7bd1e3739